requirements.txt
aioredis==1.2.0
app-version==1.0.1
apt-clone==0.2.1
apturl==0.5.2
asgi-redis==1.4.3
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
authorizenet==1.0.14
autobahn==19.5.1
Automat==0.7.0
backcall==0.1.0
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.1.6
beautifulsoup4==4.6.0
blinker==1.3
Brlapi==0.6.6
cached-property==1.5.1
certifi==2019.3.9
cffi==1.11.5
channel==0.1.0
channels==2.2.0
channels-api==0.4.0
channels-redis==2.3.3
chardet==3.0.4
colander==1.5.1
colorama==0.3.9
command-not-found==0.3
configglue==1.1.2
configobj==5.0.6
constantly==15.1.0
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
cupshelpers==1.0
daphne==2.3.0
dataclasses==0.6
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
dill==0.2.9
Django==2.2.2
django-after-response==0.2.2
django-appconf==1.0.2
django-authorize==0.1.3
django-background-task==0.1.8
django-background-tasks==1.2.0
django-braces==1.13.0
django-channels==0.7.0
django-ckeditor==5.6.1
django-common-helpers==0.9.1
django-compat==1.0.15
django-cors-headers==3.0.2
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-extensions==2.1.6
django-filter==1.0.4
django-js-asset==1.2.2
django-jsonfield==1.2.0
django-memoize==2.2.0
django-notify==1.1.2
django-notify-x==0.1.9
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.2
django-rest-framework-mongoengine==3.3.1
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.2
django-simple-email-confirmation==0.22
django-simple-email-verification==0.1.2
django-stubs==0.11.1
django-user-accounts==2.0.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
docker==2.7.0
docker-compose==1.12.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.13
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
enum34==1.1.6
feedparser==5.1.3
funcsigs==1.0.2
future==0.17.1
gunicorn==19.9.0
hiredis==1.0.0
html5lib==1.0.1
httplib2==0.9.1
hyperlink==19.0.0
idna==2.8
img==2.5
incremental==17.5.0
iniparse==0.4
ipaddress==1.0.22
ipython==7.4.0
ipython-genutils==0.2.0
iso8601==0.1.12
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonfield==2.0.2
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
lockfile==0.12.2
louis==3.5.0
lxml==4.1.1
macaroonbakery==1.1.3
Mako==1.0.3
Markdown==2.6.9
MarkupSafe==1.1.0
mock==2.0.0
msgpack==0.6.1
msgpack-python==0.5.6
mypy==0.700
mypy-extensions==0.4.1
nemo-emblems==4.0.2
netifaces==0.10.4
nose==1.3.7
notify==0.3.1
oauth2-provider==0.0
oauthlib==3.0.1
olefile==0.44
onboard==1.4.1
openapi-codec==1.3.2
Paginator==0.5.1
PAM==0.4.2
paramiko==2.4.0
parse==1.12.0
parso==0.4.0
pathlib==1.0.1
payeezy==1.0.10
pbr==3.1.1
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
piston-mini-client==0.7.5
prompt-toolkit==2.0.9
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.7.3.1
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pyasn1==0.4.5
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.3.1
pygobject==3.26.1
PyHamcrest==1.9.0
PyICU==1.9.8
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-social-auth==0.3.6
python-xapp==1.4.0
python-xlib==0.20
python3-openid==3.1.0
pytz==2019.1
PyXB==1.2.5
pyxdg==0.25
PyYAML==3.13
pyzmq==18.0.1
redis==2.10.6
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
route==0.1.2
sessioninstaller==0.0.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
texttable==0.8.8
tornado==6.0.2
traitlets==4.3.2
translationstring==1.3
Twisted==19.2.0
txaio==18.8.1
typed-ast==1.3.1
typing-extensions==3.7.2
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
whitenoise==4.1.2
xkit==0.0.0
zope.interface==4.6.0
